changes new strategy to separate file
